requirements.txt
pyspark==3.5.2
delta-spark==3.2.0
faker==28.0.0
pyarrow==16.1.0
pandas==2.2.2
plotly==5.22.0
streamlit==1.37.0
python-dotenv==1.0.1
